changed ban for posting link to kick

+++ src/bot.ts
@@ -119,16 +119,16 @@ bot.on('guildMemberAdd', (member) => {
 
 const link_regex = new RegExp(/^(http:\/\/www\.|https:\/\/www\.|http:\/\/|https:\/\/)?[a-z0-9]+([\-\.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$/, 'gm');
 const checkSpam = async (msg: Discord.Message) => {
-  if (!msg.guild) return;
+  if (!msg.guild || msg.guild.id !== servers('main').id) return;
   const index = newMembers.indexOf(msg.author.id);
   if (index < 0) return;
 
   if (!msg.content.includes('chaoticrecode') && link_regex.test(msg.content)) {
-    if (msg.member.bannable) {
-      await msg.member.ban({ reason: 'Posted link as first message. Typically spam bot behavior.' })
+    if (msg.member.kickable) {
+      await msg.member.kick('Posted link as first message. Typically spam bot behavior.')
       .then(async () => {
         const channel = bot.channels.get(servers('main').channel('staff')) as Channel;
-        return await channel.send(`Banned Spam: ${msg.member.displayName}\nContent: ||${msg.content}||`);
+        return await channel.send(`Kicked suspected spam: ${msg.member.displayName}\nContent: ||${msg.content}||`);
       })
       .then(() => { if (msg.deletable) msg.delete(); });
     }
